integration: move test_batch to batch.rs
Test `test_batch()` is extracted into `statements/batch.rs`.
